VR App Created for BFCI as a Graduation Project
Medical training has remained largely unchanged for hundreds of years despite changes in the tools and techniques used to practice medicine. While the traditional master–apprentice training method has produced and is continuing to produce highly skilled medical practitioners, there has been a push to change the methods of delivering medical training in response to the rapid technological advances that are shaping health-care provision worldwide.
For many years medical students have learned anatomy on dead bodies in cadaver labs, students dissect, touch, rotate, and explore organs in hands-on experiences. Unfortunately, these experiences are out of reach for most of us. Cadaver labs are expensive to run, cadavers are in limited supply and crowded most of the times. students have studied anatomy and Theoretically using pen and paper this is not nearly as effective as the hands-on learning experiences.
 	As the demands on medical practitioners increase, the pressure is increasing to find effective alternatives to traditional teaching practices that reduce reliance on teaching provided by expert mentors. VR medical simulation is being demonstrated as an important component of teaching reforms.it can improve training outcomes and reduce the negative impacts associated with some contemporary teaching practices.